pyproject.toml
[tool.poetry]
name = "chatgpt-client"
version = "0.1.0"
description = "Client for ChatGPT."
authors = ["Olivier Bénard"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^25.1.0"
mypy = "^1.14.1"
pylint = "^3.3.4"
pytest = "^8.3.4"
dynaconf = "^3.2.7"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
